requirements.txt
sym-api-client-python>=0.1.16
urllib3<1.25,>=1.21.
requests-pkcs12==1.4
requests-toolbelt==0.9.1
os
codecs
json